+++ src/PowerAppsTestEngineWrapper/InputOptions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-namespace PowerAppsTestEngine
+namespace PowerAppsTestEngineWrapper
 {
     public class InputOptions
     {
